+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Change Log
 ## [v1.4.0] - 2022-08-12
 ### Fixed
+- Logger now sets the level appropriately
 - Removed output codec selection fom `settings.py`. The ability to select between codecs was thought to be justified by different codecs available on different devices. Instead, selecting the fourcc encoder `mp4v` works on all devices which support h264 encoding. No need for this setting.
 - Output stage bug fixed in `comms.py` for moving temporary file to output directory
 - The YUV file format modification is a breaking change to the `yuv_buf_to_bgr_array` within `imdecode.py` 
@@ -8,6 +9,7 @@
 - The `h264_to_mp4` method within `imdecode.py` is updated to use the `-r` argument when calling `ffmpeg`. This is in the place of the `-framerate` flag which is depreciated. 
 
 ### Added 
+- Some logging messages added in `remember_from_disk.py`, changed verbosity of log messages in `filtering.py` and `event.py`
 - Within `__main__.py` a CLI is written to allow the passing of filenames, appropriate argument parsing is included to ensure a `.mp4` file is passed and that file exists. If no file is passed, the Vid2Frames library is not imported, thus DynAIkonTrap can function as normal if vid2frames is not installed. 
 - Within `comms.py`, the functions `_read_events_to_image` and `_read_events_to_video` are modified to work with `.mp4` files as well as `.h264`. This is a basic change implemented using glob. Allows for easier integration with `Vid2Frames`
 - `imdecode.py` now contains an additional function, `bgr_array_to_yuv_buf` which can generate a YUV420 buffer in an equivalent format to that produced by the camera hardware. This allows image frames to be translated into YUV buffers. Frame data from the legacy pipeline may now be linked up with the LOW_POWERED pipeline.